chore: forge init

+++ contracts/interfaces/IRaritySocietyDAOToken.sol
@@ -4,6 +4,6 @@ interface IRaritySocietyDAOToken {
 
     function getPriorVotes(address account, uint blockNumber) external view returns (uint32);
 
-    function totalSupply() external view returns (uint32);
+    function totalSupply() external view returns (uint256);
 
 }
